Phase 1: rename CJS-only build/generator files to .cjs
Rename scripts/, rosidl_gen/, rosidl_gen/templates/, rosidl_parser/, and
rostsd_gen/ modules from .js to .cjs ahead of the ESM migration (issue #1358).
These files run synchronously under Node (gyp, postinstall, message
generation) and stay CommonJS, so they must be explicitly .cjs once
"type": "module" is introduced in Phase 3.

Update all references: package.json scripts/bin, binding.gyp, .npmignore,
eslint globs, internal requires, the worker spawn path, the extensionless
py_utils require, message-template emitted require paths, and external
requires in lib/, index.js, tests, and examples.

Verified with a native rebuild, full message regeneration (35 packages),
and a runtime load smoke test under ROS jazzy.
