+++ lib/message_introspector.js
@@ -15,7 +15,7 @@
 'use strict';
 
 const loader = require('./interface_loader.js');
-const { toPlainObject } = require('../rosidl_gen/message_translator.js');
+const { toPlainObject } = require('../rosidl_gen/message_translator.cjs');
 const { TypeValidationError } = require('./errors.js');
 
 /**
