+++ package.json
@@ -35,30 +35,30 @@
     "build:dev": "node-gyp -j 16 build --debug",
     "rebuild": "npm run clean && node-gyp -j 16 rebuild",
     "rebuild:dev": "npm run clean && node-gyp -j 16 rebuild --debug",
-    "generate-messages": "node scripts/generate_messages.js",
-    "generate-messages-idl": "node scripts/generate_messages.js --idl",
-    "generate-messages:dev": "node scripts/generate_messages.js --debug",
-    "generate-tsd-messages": "node scripts/generate_tsd.js",
+    "generate-messages": "node scripts/generate_messages.cjs",
+    "generate-messages-idl": "node scripts/generate_messages.cjs --idl",
+    "generate-messages:dev": "node scripts/generate_messages.cjs --debug",
+    "generate-tsd-messages": "node scripts/generate_tsd.cjs",
     "clean": "node-gyp clean && npx rimraf ./generated",
-    "install": "node scripts/install.js",
+    "install": "node scripts/install.cjs",
     "postinstall": "npm run generate-messages",
     "docs": "make -C tools/jsdoc",
     "docs:gh-pages": "node tools/jsdoc/regenerate-published-docs.js --branch origin/gh-pages --preserve-published",
     "docs:gh-pages:full": "node tools/jsdoc/regenerate-published-docs.js --branch origin/gh-pages --full-rebuild",
-    "test": "nyc node --expose-gc ./scripts/run_test.js && tsd && npm install --no-save electron && node test/electron/run_test.js",
-    "test-idl": "nyc node --expose-gc ./scripts/run_test.js --idl",
-    "lint": "eslint && node ./scripts/cpplint.js",
+    "test": "nyc node --expose-gc ./scripts/run_test.cjs && tsd && npm install --no-save electron && node test/electron/run_test.js",
+    "test-idl": "nyc node --expose-gc ./scripts/run_test.cjs --idl",
+    "lint": "eslint && node ./scripts/cpplint.cjs",
     "test:asan": "bash scripts/run_asan_test.sh",
-    "format": "clang-format -i -style=file ./src/*.cpp ./src/*.h && npx --yes prettier --write \"{lib,rosidl_gen,rostsd_gen,rosidl_parser,types,example,test,scripts,benchmark,rostsd_gen}/**/*.{js,md,ts}\" ./*.{js,md,ts}",
+    "format": "clang-format -i -style=file ./src/*.cpp ./src/*.h && npx --yes prettier --write \"{lib,rosidl_gen,rostsd_gen,rosidl_parser,types,example,test,scripts,benchmark,rostsd_gen}/**/*.{js,cjs,md,ts}\" ./*.{js,md,ts}",
     "prepare": "husky",
     "coverage": "cat ./coverage/lcov.info | ./node_modules/coveralls/bin/coveralls.js",
     "prebuild:node": "prebuildify --napi --strip --name node --target 20.20.2",
     "prebuild:electron": "prebuildify --napi --strip --name electron --target electron@34.0.0",
-    "prebuild": "npm run prebuild:node && npm run prebuild:electron && node scripts/tag_prebuilds.js",
+    "prebuild": "npm run prebuild:node && npm run prebuild:electron && node scripts/tag_prebuilds.cjs",
     "rosocket": "node ./rosocket/cli.js"
   },
   "bin": {
-    "generate-ros-messages": "./scripts/generate_messages.js",
+    "generate-ros-messages": "./scripts/generate_messages.cjs",
     "rclnodejs-web": "./bin/rclnodejs-web.js",
     "rosocket": "./rosocket/cli.js"
   },
