+++ scripts/config.cjs
