+++ scripts/install.cjs
