+++ test/test-security-related.js
@@ -17,7 +17,7 @@
 const assert = require('assert');
 const rclnodejs = require('../index.js');
 const assertThrowsError = require('./utils.js').assertThrowsError;
-const translator = require('../rosidl_gen/message_translator.js');
+const translator = require('../rosidl_gen/message_translator.cjs');
 const arrayGen = require('./array_generator.js');
 
 describe('Destroying non-existent objects testing', function () {
